requirements.txt
azure-cognitiveservices-speech==1.41.1
requests
wyoming==1.6.0
